Expirement with queue to try and render every states cities after map renders
