+++ package.json
@@ -23,6 +23,7 @@
     "knex-seed-file": "^0.3.1",
     "md5": "^2.2.1",
     "nodemon": "^1.11.0",
-    "pg": "^6.1.5"
+    "pg": "^6.1.5",
+    "require": "^2.4.20"
   }
 }
