+++ public/index.html
@@ -3,6 +3,7 @@
 <head>
   <title>Fatal Police Shootings in the U.S.</title>
   <link rel="stylesheet" type="text/css" href="style/style.css">
+  <script src='js/helpers.js'></script>
   <script src='js/city-finder.js'></script>
   <script src="https://d3js.org/d3.v4.js"></script>
   <script src="https://d3js.org/topojson.v2.min.js"></script>
