+++ server.js
@@ -1,8 +1,8 @@
 const express = require('express')
 const bodyParser = require('body-parser')
-const cors = require('cors');
-const fs = require('fs');
-const path = require('path');
+const cors = require('cors')
+const fs = require('fs')
+const path = require('path')
 const app = express()
 
 const environment = process.env.NODE_ENV || 'development'
